requirements.txt
pandas
numpy
faker
python-dateutil
pyarrow
sqlalchemy
psycopg2-binary
apache-airflow
apache-airflow-providers-postgres
apache-airflow-providers-amazon
boto3
requests
tqdm
matplotlib
scikit-learn
nltk
google-cloud-bigquery
google-cloud-storage